Feature/admin portal (#29)
* Added : Basic admin portal structure

* Refractor: Cleaned up code to remove redundancy 🧹 (#28)

* Refractor: Cleaned up code to remove redundancy 🧹

* Fixed: alignement issue with one of the cards

* Refractor: Sorted components in sub-directory

* Refractor: Cleaned up Footer Component

* Refractor: Cleaned up Footer and Loader Components

* Added : Ability to add our own projects

---------

+++ src/pages/admin/page.tsx
@@ -1,17 +1,14 @@
-import { Button } from '@/components/ui/button';
+import Admin from '@/components/Admin';
+import WrongPage from '@/components/WrongPage';
 import { client } from '@/lib/supabase/supabase';
-import { useRouter } from 'next/router';
 import React, { useEffect, useState } from 'react'
 
 const AdminPage = () => {
-    const router = useRouter();
     const [validUser, setValidUser] = useState(false);
+
     useEffect(() => {
         const checkSession = async () => {
             const { data: { session } } = await client.auth.getSession();
-
-            console.log(session);
-            console.log(new Date().getTime());
 
             if (session) {
                 setValidUser(true);
@@ -19,35 +16,21 @@ const AdminPage = () => {
             }
             return;
         }
-
+        client.auth.onAuthStateChange((event, session) => {
+            if (event === "SIGNED_OUT") {
+                localStorage.clear();
+            }
+        })
         checkSession();
     }, []);
-
-    const handleLogout = () => {
-        localStorage.clear();
-        router.push("/");
-    }
 
     return (
         <>
             {
                 validUser ? (
-                    <div>
-                        AdminPage
-                        <Button onClick={handleLogout}>
-                            Logout
-                        </Button>
-                    </div>
+                    <Admin />
                 ) : (
-                    <div>
-                        <p>
-                            Oops!! You're not supposed to be here!!
-                        </p>
-                        <Button onClick={handleLogout}>
-                            Return To Home
-                        </Button>
-                    </div>
-
+                    <WrongPage />
                 )
             }
         </>
